Development - Dynamic basic simul and scenarios ready (#182)
* FIX: not needed notification on model load/perspective change
FIX: Expressions for OTRel of soft goals

* FIX: Configuration operation working with new attributes.

* FIX: Define all soft goals related attributes with the same domain
(0..4)

* FIX: Order corrected for new simulation
FIX: Broken models avoided when a clone is lost and the clonned element
has relations. The element must be deleted and the relations created
again, but the model is not broken.

* FIX: constructor for RIGHTCONCEPTVARIABLE with correct semantic concept
FIX: RIGHTVARIABLE replaced by RIGHTCONCEPTVARIABLE for refas opers meta
model
FIX: expressions for soft goals old implementation fixed. New
implementation still incomplete

* FIX: sg to sg otrel working for dynamic operations (not optimal, not
suitable for MUTEX or RANGE): works for AND and OR

* FIX: use of OTRel and PWRel types visually defined. DisplayName and
Panel name not working.

* NEWFUN: Include tooltip for attribute's labels

* FIX: some tooltips included

* FIX: Required Level working for sg in static simulation. Included in
domain values for ClaimExpLevel

* FIX: duplicated relation type removed

* IMP: Many errors displayed on message area tab.
FIX: simulation for scenarios working without sorting expressions.

* FIXES: dynamic opers

* FIX: dynamic simulation working without label order

* FIX: laber order working for basic dynamic simulation

* FIX

* FIX: edition of syntax meta model elements

* FIX: pairwise relation for claim to sg fixed for old operations.
Missing class from old implementation

* FIX: domain for enum variables in static simulation
IMP: hints for attributes
DOC: future method to abort solver execution

* FIX: missing class

* IMP: Old relation objects for relation types and methods removed

* FIX: relations types display using with DispValue visually defined

* fix: Duplicated condition for dynamic relations types removed.

* IMP: All Syntax classes reduced to SyntaxElement
NEWFUN: float domain partially supported.

* IMP: Operation classes for the model reduced to OpersConcept class. Not
needed classes removed.

* IMP: Additional attributes to support custom filtered domains

* FIX: Dynamic opers using temporal filters for domains.

* FIX:regex for float changed

* FIX: monitoring frezze when the configuration is the same

* FIX: reltaion types for some relations
FIX: mape working for same fconfiguration

* IMP: Dynamic basic expression working with static domain filtering for
variables and soft goals

* NEWFUN: Dynamic operations using dynamic filters for soft goals and
variables

* FIX: Nofitication of no solution for dynamic opers fixed.

+++ com.variamos.compiler/src/com/variamos/compiler/prologEditors/Hlcl2SWIProlog.java
@@ -291,8 +291,8 @@ public class Hlcl2SWIProlog extends Hlcl2Prolog implements SWIPrologSymbols {
 	 */
 	private StringBuffer getRangeDomain(Domain domain, String id) {
 		StringBuffer domainString = new StringBuffer();
-		Integer lowerValue = ((RangeDomain) domain).getLowerValue();
-		Integer upperValue = ((RangeDomain) domain).getUpperValue();
+		Integer lowerValue = (int) ((RangeDomain) domain).getLowerValue();
+		Integer upperValue = (int) ((RangeDomain) domain).getUpperValue();
 		if (id != null) {
 			domainString.append(id);
 			domainString.append(IN);
